Make IANA PEN download configurable - fix uninitalized DOWNLOAD variable
Currently if you do not have wget and curl requirement met, you get
this error:
[   93s] configure: WARNING: ** Neither wget nor curl could be found.
[   93s] configure: WARNING: ** IANA PEN database will not be installed by `make install` !
[   93s] configure: WARNING: ** Download is:
[   93s] configure: WARNING:
...
[  104s] configure: error: conditional "DOWNLOAD" was never defined.
[  104s] Usually this means the macro was only invoked conditionally.
[  104s] error: Bad exit status from /var/tmp/rpm-tmp.TYnvu5 (%build)

Internet download is restricted in most build environments.
So there must be a knob to enable/disable IANA PEN database download.
For security reasons and as a good manner for open source tools, the internet
download is by default set to off.

This patch initializes all needed variables and also introduces to make the
IANA PEN internet download configurable.

./configure
then has this additional feature:

  --enable-iana-download  Download IANA PEN database [default=no]

Depending on whether it has explicitly been enabled this additional output
is shown after build env is successfully set up via ./configure:

  Download IANA PEN database   : yes
  IANA PEN database URL        : http://www.iana.org/assignments/enterprise-numbers

The URL is unfortunately hardcoded in the message. I couldn't find a quick
way to show the IANA_PEN_URL variable there, so if this is ever changed (it is
not configurable right now, but maybe with a follow up patch in the future),
it has to be changed in the help string as well.